+++ src/main/java/com/colvir/shop/config/Config.java
@@ -1,8 +1,10 @@
 package com.colvir.shop.config;
 
 import com.colvir.shop.repository.CatalogRepository;
+import com.colvir.shop.repository.CategoryRepository;
 import com.colvir.shop.repository.impl.CatalogRepositoryMemoryImpl;
 import com.colvir.shop.repository.impl.CatalogRepositoryPostgresImpl;
+import com.colvir.shop.repository.impl.CategoryRepositoryMemoryImpl;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -22,6 +24,16 @@ public class Config {
         return switch (repositoryMode) {
             case "memory" -> new CatalogRepositoryMemoryImpl();
             case "postgres" -> new CatalogRepositoryPostgresImpl();
+            default ->
+                    throw new RuntimeException(String.format("Режим репозитория %s не поддерживается.", repositoryMode));
+        };
+    }
+
+    @Bean
+    public CategoryRepository getCategoryRepository () {
+        return switch (repositoryMode) {
+            case "memory" -> new CategoryRepositoryMemoryImpl();
+//            case "postgres" -> new CategoryRepositoryPostgresImpl(getJdbcTemplate()); TODO. Реализовать имплементацию
             default ->
                     throw new RuntimeException(String.format("Режим репозитория %s не поддерживается.", repositoryMode));
         };
